Copy platform "ps" parser into ZenPack.
Depending on the Zenoss version the platform "ps" COMMAND datasource
parser might have converted process memory usage from kbytes to bytes.
If it was already converted to bytes by the parser (good idea) then we
don't need an RPN on our graphpoints and aliases, but if it wasn't
converted, we do need those RPNs.

The only sane thing to do is bring the "ps" parser into the ZenPack so
we can be sure not to erroneously apply the calculation the wrong number
of times.

Fixes ZEN-21166.