Fix "Component Graphs" for custom getGraphObjects
This ZenPack makes heavy use of overridding component.getGraphObjects()
to show useful graphs from related contexts for many component types.
Currently this isn't well-supported by the platform. All of the
other-contexted graph definitions will appear as choices in the graph
selection when the component type is selected. If any of these
other-contexted graphs are chosen, no data will be displayed on the
graphs.

This monkeypatch fixes the platform to filter out other-contexted graph
definitions, but only for Linux devices.

Fixes ZEN-22254.