Fix lvsstatus parser now that LV ids are prefixed
When the lvsstatus parser was written, LogicalVolume and SnapshotVolume
ids were just "<vgname_lvname>" which matched the output of the lvs
command being run for this parser. Now we're prefixing LV ids with "lv-"
so the match no longer works.

This change simplies the command being run by removing awk pre-parsing
and letting the lvsstatus parser do all the parsing work by specifically
matching the VG name and LV name in the output columns with the VG name
and LV name of each LogicalVolume and SnapshotVolume.

Fixes ZEN-22301.