Restore aliases used by performance reports
Restored the following datapoint aliases:

* usedFilesystemSpace_bytes (used by Filesystem Util Report)
* loadAverage5min (used by CPU Utilization report)
* inputOctets__bytes (used by Interface Utilization report)
* outputOctets__bytes (used by Interface Utilization report)
* memoryBuffered__bytes (used by Memory Utilization report)
* memoryCached__bytes (used by Memory Utilization report)
* memoryAvailable__bytes (used by Memory Utilization report)
* memoryUsed__pct (used by Memory Utilization report)

There's still a big problem with all of these performance reports. It
appears that the datapoint alias RPN formulas are not being calculated
at all. So the while this fix makes numbers appear, most of them are
completely wrong. This appears to be a platform issue in
metric-service/CentralQuery for which I'll open a separate defect.

Fixes ZEN-22310.