Workaround "All on Same Graph?" issue for PV/VG
There are a couple of issues (ZEN-22411 and ZEN-22413)  with Component
Graphs in Zenoss 5.1.1 that cause them not work when a graph point's RPN
refers to another graph point. We used this feature for PhysicalVolume
and VolumeGroup Utilization graphs.

To workaround the issue we use a TALES expression to get the current
size of the given PV/VG instead. This seems to work, but is not as good
in the long run because it doesn't accomodate for the size of the volume
changing over time as well.

Fixes ZEN-22254.