Revert status to be defined only by /Status/Ping
The default definition of "status" in Zenoss is that the presence of
error or critical events in the /Status/Ping event class means down, and
the absence of such events means up.

A previous fix went in for ZEN-20498 to change the meaning of status for
/Server/SSH/Linux devices to continue using /Status/Ping to identify
down devices, but to also use /Cmd/Fail where zencommand creates events
cause by SSH failures. The idea was that if we're mostly monitoring a
device using SSH that the device is down if we can't. This turned out to
not be true at all. There are all sorts of /Cmd/Fail events generated by
zencommand that don't indicate the device is down. It could just be a
wrong username, or no username configured.

Since we don't have control over the events sent by zencommand when SSH
fails in various ways within this ZenPack, we can't change the meaning
of down to include those events in this ZenPack.

Unfixes ZEN-20498.