Prevent thresholds on NICs with unknown speed
Previously we were assuming interfaces without a known speed were
10Mbps. This seems like a really bad guess, and as of 29d94c9 we are
much better able to determine the speed of interfaces.

This change to assume a 1Tbps interface speed in the absence of a known
speed is intended to prevent any threshold violations on interfaces for
which we still don't know the speed. This would apply to special
interface types like bridges. For bridges we're getting usage and speed
from the member interfaces anyway.