Fix NewDeviceType migration
The NewDeviceType migration optimization (8a311e2) seems to fail in
changing the class of Linux devices from Device
(Products.ZenModel.Device.Device) to LinuxDevice
(ZenPacks.zenoss.LinuxMonitor.LinuxDevice.LinuxDevice) due to ZODB not
storing the change to the device's __class__ attribute. It appears that
the change is successful, but upon reloading the Zope client, __class__
is reset back to its original value.

This is apparently a known behavior of ZODB, but there is a workaround.
This commit uses that workaround to also replace the object's persistent
reference and cause the __class__ change to be permanent.

Fixes ZEN-24293.