Fix testPlugins and TestDVI unit tests
No problems in the code being tested. Just problems in the unit tests.

- testPlugins failed due to a slight change in the command being run.
- TestDVI was expecting DynamicView and Impact to be installed.

These tests have already been fixed on the develop branch. Fixing them
again in a hotfix so we can keep Zenoss 5.2 tests passing with a stable
version of LinuxMonitor.

ZEN-24653 will only be fixed once LinuxMonitor 2.0.5 is released with
this fix.