Fix inconsistencies with our data and df's output
Attempting to calculate our own capacity value that matched the
"Capacity" column in Linux's df output turns out to be pretty much
impossible due to the various per-filesystem special reservations that
can make used + available != total.

This change takes what I hope is the simplest and best approach to
addressing this complication. We'll just collect and use the "Capacity"
column from df's output in all of the places where people expect that
value.

- Add percentUsed datapoint to filesystems
- Use percentUsed datapoint in Utilization graph
- Use percentUsed datapoint in default "90 percent used" threshold
- Use percentUsed in /Perf/Filesystem event transform

See ZPS-1545 for more detail on the problems that this change addresses.

Fixes ZPS-1545.