Fix NoneType flare when clicking Storage Device
The custom renderer used for the filesystem "Storage Device" column
assumed that the storage device would be a component on the same device
such as a HardDisk or LogicalVolume component. In cases such as this
where the storage device is a NetApp FileSystem mounted via NFS, this
assumption breaks.

Updated the custom renderer to use a same-device link if the storage
device is on the same device as the filesystem, and to use a regular
link otherwise.

Fixes ZPS-1822.