Fix zencommand warnings monitoring NFS filesystems
Because the commandTemplate on these COMMAND datasources didn't begin
with a / (forward slash), zencommand would prefix the entire command
with the device's zCommandPath value, which is $ZENHOME/libexec by
default. This resulted in zencommand log messages like the following.

    WARNING zen.SshClient: <ip-address> channel 450 The command $ZENHOME/libexec/export PATH=$PATH:/bin:/sbin:/usr/bin:/usr/sbin; if command -v timeout > /dev/null 2>&1; then timeout 30 /usr/bin/env df -ikP; else /usr/bin/env df -ikP; fi returned stderr data (1) from the device: bash: /libexec/export: No such file or directory

It appears this didn't prevent monitoring from working because only the
export resulted in an error, and the other commands continued to
execute, and their output was parsed.

Fixes ZPS-1823.