Use MemAvailable instead of MemFree for Used %
MemAvailable exists in newer kernels, and the difference between
MemTotal and MemAvailable is a more useful indicator of how much memory
is actively being used on a a system than the difference between
MemTotal and MemFree.

https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=34e431b0ae398fc54ea69ff85ec700722c9da773

The intention of 05269b2 was to do this, but the combination of the mem
parser's code, and the lack of a MemAvailable datapoint on the mem
datasource resulted in MemFree still being used. This change corrects
that by not requiring a MemAvailable datapoint to exist.

Fixes ZPS-1144.