Fix 0% util on Windows filesystem threshold events
The problem was in the following code.

    usedPercent = (usedBytes / totalBytes) * 100.0

The dividend and divisor were both integers, so all precision was lost.
This was corrected to the following in both places.

    usedPercent = (usedBytes / float(totalBytes)) * 100.0

Having either the divident or divisor be a float results in floating
point division.

Note that another problem with this transform is also fixed by these
changes. The transform installation/update logic was moved into the
ZenPack's install method instead of being in a migrate step. This makes
sure that even new installations get the proper transform.

Fixes ZPS-1844.