Prevent monitoring of potentially wrong services
In LinuxMonitor 2.3.0 we began monitoring Linux services. There were
some problems in how we decided which services to model that resulted in
services that weren't supposed to be running (disabled in some way)
being modeled, and then being monitored. These services would flood the
system with events about services that were supposed to be down being
down.

We corrected this in LinuxMonitor 2.3.1, but in the process we added the
sudo execution of new commands to the cmd.linux.os_service modeler
plugin. This has had the result of making modeling fail for environments
where the sudo configuration was locked down to the previous set of sudo
commands. This modeling failure causes the erroneously-modeled "supposed
to be down" services to remain modeled, which results in monitoring
continuing, and the event spam for down services continuing.

One solution for this is for the customer to update their sudo
configurations to allow the new commands to be run according to the
documentation, but this isn't always an easy thing for customers to
change across a large environment.

This change attempts to address the problem by adding a
"modeler_version" property to LinuxService components. This allows us to
disable monitoring of services until such time as they're successfully
modeled using the new, fixed, cmd.linux.os_service modeler plugin.

The customer will have a modeling event on each Linux device where the
sudo configuration needs to be updated as a way of bringing light to the
correction that needs to be made.

Fixes ZPS-4334.