consolidate impact edge providers
See IMP-725 for the Impact side of this change, and ZEN-31678 for the
DynamicView side. The idea is to consolidate all impact edge creation
for LinuxMonitor devices and components into this ZenPack in the form of
DynamicView relationships.

Fixes ZPS-5654.

(cherry picked from commit d92522ca329129e3356fb79a7caef51141d61c78)
Signed-off-by: Chet Luther <chet.luther@gmail.com>