fix disks being impacted by random objects.
When modeling "disk_ids" that can be used to correlate disks to their
underlying storage provider, we try to pick serial numbers out. However,
there was a particularly greedy case that looked at Ubuntu swap disks
like "disk-qa--ubuntu--12-swap_1" and picked off the last "1" as a
serial number. We then search for any components in the system that have
"1" as a keyword and say that they impact us. This is definitely not a
good thing.

This change attempts to keep the behavior as close as possible to what
was originally intended while avoiding overly-simple keywords by
expecting serial numbers to have over 12 characters. The test data we
have for this feature shows only serial numbers 18 or more characters.

Fixes ZPS-5792.