fix UCS disk impact regression in c24bbc0
In commit c24bbc0 I changed the modeler plugin to require that serial
numbers in disk IDs must be longer than 12 characters. It turns out that
this was too long. A drive serial number of "9XG56VGD" was found. So I'm
lowering that limit to include serial numbers longer than 7 characters.

Fixes ZPS-5828.